Everything done except results and docstrings
