+++ src/LogisticRegression/LogisticRegression.py
@@ -1,96 +1,18 @@
-import numpy as np
 import argparse
-import matplotlib.pyplot as plt
-import pandas as pd
 from LogisticRegressionModel import LogisticRegressionModel
-import random
-
-
-def write_beta(output, beta_to_save, feature_names):
-    with open(output, 'w') as f:
-        for val, label in zip(beta_to_save, feature_names):
-            f.write('(%s,%f)\n' % (label, val))
-
-
-def read_features_pd(input_file: str) -> pd.DataFrame:
-    data = pd.read_csv(input_file)
-    data.drop(columns=['Diabetes_binary'], inplace=True)
-    return data
-
-
-def read_features_np(input_file: str) -> np.array:
-    return np.array(read_features_pd(input_file))
-
-
-def read_labels_pd(input_file: str) -> pd.DataFrame:
-    return pd.read_csv(input_file)['Diabetes_binary']
-
-
-def read_labels_np(input_file: str) -> np.array:
-    return np.array(read_labels_pd(input_file))
-
-
-def get_all_features(data: pd.DataFrame) -> list:
-    return list(data.columns)
-
-
-def random_oversample(features, labels):
-    pos_feats = [x for i, x in enumerate(features) if labels[i] == 1]
-    neg_feats = [x for i, x in enumerate(features) if labels[i] == 0]
-    num_pos = len(pos_feats)
-    num_neg = len(neg_feats)
-    num_less = abs(num_pos - num_neg)
-    choice_arr = pos_feats if num_pos < num_neg else neg_feats
-    choice_label = 1 if num_pos < num_neg else 0
-    for i in range(num_less):
-        features.append(random.choice(choice_arr))
-        labels.append(choice_label)
-    return np.array(features), np.array(labels)
-
-
-# https://stackoverflow.com/questions/4601373/better-way-to-shuffle-two-numpy-arrays-in-unison
-def unison_shuffled_copies(a, b):
-    assert len(a) == len(b)
-    p = np.random.permutation(len(a))
-    return a[p], b[p]
-
-
-def create_data_plots(k, data, label=""):
-    plt.plot([i for i in range(1, k)], data, label=label)
-    plt.xlabel("Number of Iterations")
-    plt.ylabel(label)
-    plt.title(f'{label} vs Number of Iterations')
-    plt.savefig(f'./LogisticRegression/results/{label}_fig_unparallelized.png')
-
-
-def create_metric_plots(k, metrics, label=""):
-    iters = [x for x in range(1, k)]
-    accs = [x[0] for x in metrics]
-    pres = [x[1] for x in metrics]
-    recs = [x[2] for x in metrics]
-    plt.plot(iters, accs, label="Accuracy")
-    plt.plot(iters, pres, label="Precision")
-    plt.plot(iters, recs, label="Recall")
-    plt.legend(loc='upper right')
-    plt.xlabel("Number of Iterations")
-    plt.ylabel("Metric Value")
-    plt.title(f'{label} Metrics vs Number of Iterations')
-    plt.savefig(f'./LogisticRegression/results/metrics_fig_unparallelized.png')
+from LogRegUtils import *
 
 
 if __name__ == "__main__":
     # Set up argument parser and associated flags
-    parser = argparse.ArgumentParser(description="Logistic Regression.",
-                                     formatter_class=argparse.ArgumentDefaultsHelpFormatter)
-    parser.add_argument('--data', default='../data/Clean_data.csv',
-                        help='Input file containing all features and labels, used to train a logistic model')
+    parser = argparse.ArgumentParser(description="Logistic Regression.", formatter_class=argparse.ArgumentDefaultsHelpFormatter)
+    parser.add_argument('--data', default='../data/Clean_data.csv', help='Input file containing all features and labels, used to train a logistic model')
     parser.add_argument('--beta', default='./LogisticRegression/results/beta.txt', help='File where beta is stored')
-    parser.add_argument('--split', type=float, default=0.8,
-                        help='Test/Training split. Percentage of data to be used for training')
+    parser.add_argument('--split', type=float, default=0.8, help='Test/Training split. Percentage of data to be used for training')
     parser.add_argument('--lam', type=float, default=0.0, help="Regularization parameter λ")
     parser.add_argument('--max_iter', type=int, default=2, help='Maximum number of iterations')
-    parser.add_argument('--eps', type=float, default=0.1,
-                        help='ε-tolerance. If the l2_norm gradient is smaller than ε, gradient descent terminates.')
+    parser.add_argument('--eps', type=float, default=0.1, help='ε-tolerance. If the l2_norm gradient is smaller than ε, gradient descent terminates.')
+    parser.add_argument('--alpha', type=float, default=1e-7, help='Learning rate controlling the step size of gradient descent')
 
     args = parser.parse_args()
 
@@ -112,7 +34,7 @@ if __name__ == "__main__":
     print("Oversampling complete")
 
     feature_labels = get_all_features(features_pd)
-    model = LogisticRegressionModel(num_features=len(feature_labels))
+    model = LogisticRegressionModel(num_features=len(feature_labels), learning_rate=args.alpha)
     TRAIN_SPLIT = args.split
     END_IDX = int(len(labels_np) * TRAIN_SPLIT)
